+++ include/roq/fbs/api.hpp
@@ -5094,7 +5094,8 @@ struct OrderAck FLATBUFFERS_FINAL_CLASS : private ::flatbuffers::Table {
     VT_SIDE = 34,
     VT_ROUND_TRIP_LATENCY = 36,
     VT_TRADED_QUANTITY = 38,
-    VT_RISK_CHANGE = 40
+    VT_RISK_EXPOSURE = 40,
+    VT_RISK_EXPOSURE_CHANGE = 42
   };
   uint16_t stream_id() const {
     return GetField<uint16_t>(VT_STREAM_ID, 0);
@@ -5150,8 +5151,11 @@ struct OrderAck FLATBUFFERS_FINAL_CLASS : private ::flatbuffers::Table {
   double traded_quantity() const {
     return GetField<double>(VT_TRADED_QUANTITY, std::numeric_limits<double>::quiet_NaN());
   }
-  double risk_change() const {
-    return GetField<double>(VT_RISK_CHANGE, std::numeric_limits<double>::quiet_NaN());
+  double risk_exposure() const {
+    return GetField<double>(VT_RISK_EXPOSURE, std::numeric_limits<double>::quiet_NaN());
+  }
+  double risk_exposure_change() const {
+    return GetField<double>(VT_RISK_EXPOSURE_CHANGE, std::numeric_limits<double>::quiet_NaN());
   }
   bool Verify(::flatbuffers::Verifier &verifier) const {
     return VerifyTableStart(verifier) &&
@@ -5181,7 +5185,8 @@ struct OrderAck FLATBUFFERS_FINAL_CLASS : private ::flatbuffers::Table {
            VerifyField<uint8_t>(verifier, VT_SIDE, 1) &&
            VerifyField<int64_t>(verifier, VT_ROUND_TRIP_LATENCY, 8) &&
            VerifyField<double>(verifier, VT_TRADED_QUANTITY, 8) &&
-           VerifyField<double>(verifier, VT_RISK_CHANGE, 8) &&
+           VerifyField<double>(verifier, VT_RISK_EXPOSURE, 8) &&
+           VerifyField<double>(verifier, VT_RISK_EXPOSURE_CHANGE, 8) &&
            verifier.EndTable();
   }
 };
@@ -5244,8 +5249,11 @@ struct OrderAckBuilder {
   void add_traded_quantity(double traded_quantity) {
     fbb_.AddElement<double>(OrderAck::VT_TRADED_QUANTITY, traded_quantity, std::numeric_limits<double>::quiet_NaN());
   }
-  void add_risk_change(double risk_change) {
-    fbb_.AddElement<double>(OrderAck::VT_RISK_CHANGE, risk_change, std::numeric_limits<double>::quiet_NaN());
+  void add_risk_exposure(double risk_exposure) {
+    fbb_.AddElement<double>(OrderAck::VT_RISK_EXPOSURE, risk_exposure, std::numeric_limits<double>::quiet_NaN());
+  }
+  void add_risk_exposure_change(double risk_exposure_change) {
+    fbb_.AddElement<double>(OrderAck::VT_RISK_EXPOSURE_CHANGE, risk_exposure_change, std::numeric_limits<double>::quiet_NaN());
   }
   explicit OrderAckBuilder(::flatbuffers::FlatBufferBuilder &_fbb)
         : fbb_(_fbb) {
@@ -5278,9 +5286,11 @@ inline ::flatbuffers::Offset<OrderAck> CreateOrderAck(
     roq::fbs::Side side = roq::fbs::Side::Undefined,
     int64_t round_trip_latency = 0,
     double traded_quantity = std::numeric_limits<double>::quiet_NaN(),
-    double risk_change = std::numeric_limits<double>::quiet_NaN()) {
+    double risk_exposure = std::numeric_limits<double>::quiet_NaN(),
+    double risk_exposure_change = std::numeric_limits<double>::quiet_NaN()) {
   OrderAckBuilder builder_(_fbb);
-  builder_.add_risk_change(risk_change);
+  builder_.add_risk_exposure_change(risk_exposure_change);
+  builder_.add_risk_exposure(risk_exposure);
   builder_.add_traded_quantity(traded_quantity);
   builder_.add_round_trip_latency(round_trip_latency);
   builder_.add_version(version);
@@ -5327,7 +5337,8 @@ inline ::flatbuffers::Offset<OrderAck> CreateOrderAckDirect(
     roq::fbs::Side side = roq::fbs::Side::Undefined,
     int64_t round_trip_latency = 0,
     double traded_quantity = std::numeric_limits<double>::quiet_NaN(),
-    double risk_change = std::numeric_limits<double>::quiet_NaN()) {
+    double risk_exposure = std::numeric_limits<double>::quiet_NaN(),
+    double risk_exposure_change = std::numeric_limits<double>::quiet_NaN()) {
   auto account__ = account ? _fbb.CreateString(account) : 0;
   auto exchange__ = exchange ? _fbb.CreateString(exchange) : 0;
   auto symbol__ = symbol ? _fbb.CreateString(symbol) : 0;
@@ -5356,7 +5367,8 @@ inline ::flatbuffers::Offset<OrderAck> CreateOrderAckDirect(
       side,
       round_trip_latency,
       traded_quantity,
-      risk_change);
+      risk_exposure,
+      risk_exposure_change);
 }
 
 struct OrderUpdate FLATBUFFERS_FINAL_CLASS : private ::flatbuffers::Table {
@@ -5398,7 +5410,8 @@ struct OrderUpdate FLATBUFFERS_FINAL_CLASS : private ::flatbuffers::Table {
     VT_USER = 68,
     VT_SENDING_TIME_UTC = 70,
     VT_CLIENT_ORDER_ID = 72,
-    VT_RISK_CHANGE = 74
+    VT_RISK_EXPOSURE = 74,
+    VT_RISK_EXPOSURE_CHANGE = 76
   };
   uint16_t stream_id() const {
     return GetField<uint16_t>(VT_STREAM_ID, 0);
@@ -5505,8 +5518,11 @@ struct OrderUpdate FLATBUFFERS_FINAL_CLASS : private ::flatbuffers::Table {
   const ::flatbuffers::String *client_order_id() const {
     return GetPointer<const ::flatbuffers::String *>(VT_CLIENT_ORDER_ID);
   }
-  double risk_change() const {
-    return GetField<double>(VT_RISK_CHANGE, std::numeric_limits<double>::quiet_NaN());
+  double risk_exposure() const {
+    return GetField<double>(VT_RISK_EXPOSURE, std::numeric_limits<double>::quiet_NaN());
+  }
+  double risk_exposure_change() const {
+    return GetField<double>(VT_RISK_EXPOSURE_CHANGE, std::numeric_limits<double>::quiet_NaN());
   }
   bool Verify(::flatbuffers::Verifier &verifier) const {
     return VerifyTableStart(verifier) &&
@@ -5554,7 +5570,8 @@ struct OrderUpdate FLATBUFFERS_FINAL_CLASS : private ::flatbuffers::Table {
            VerifyField<int64_t>(verifier, VT_SENDING_TIME_UTC, 8) &&
            VerifyOffset(verifier, VT_CLIENT_ORDER_ID) &&
            verifier.VerifyString(client_order_id()) &&
-           VerifyField<double>(verifier, VT_RISK_CHANGE, 8) &&
+           VerifyField<double>(verifier, VT_RISK_EXPOSURE, 8) &&
+           VerifyField<double>(verifier, VT_RISK_EXPOSURE_CHANGE, 8) &&
            verifier.EndTable();
   }
 };
@@ -5668,8 +5685,11 @@ struct OrderUpdateBuilder {
   void add_client_order_id(::flatbuffers::Offset<::flatbuffers::String> client_order_id) {
     fbb_.AddOffset(OrderUpdate::VT_CLIENT_ORDER_ID, client_order_id);
   }
-  void add_risk_change(double risk_change) {
-    fbb_.AddElement<double>(OrderUpdate::VT_RISK_CHANGE, risk_change, std::numeric_limits<double>::quiet_NaN());
+  void add_risk_exposure(double risk_exposure) {
+    fbb_.AddElement<double>(OrderUpdate::VT_RISK_EXPOSURE, risk_exposure, std::numeric_limits<double>::quiet_NaN());
+  }
+  void add_risk_exposure_change(double risk_exposure_change) {
+    fbb_.AddElement<double>(OrderUpdate::VT_RISK_EXPOSURE_CHANGE, risk_exposure_change, std::numeric_limits<double>::quiet_NaN());
   }
   explicit OrderUpdateBuilder(::flatbuffers::FlatBufferBuilder &_fbb)
         : fbb_(_fbb) {
@@ -5719,9 +5739,11 @@ inline ::flatbuffers::Offset<OrderUpdate> CreateOrderUpdate(
     ::flatbuffers::Offset<::flatbuffers::String> user = 0,
     int64_t sending_time_utc = 0,
     ::flatbuffers::Offset<::flatbuffers::String> client_order_id = 0,
-    double risk_change = std::numeric_limits<double>::quiet_NaN()) {
+    double risk_exposure = std::numeric_limits<double>::quiet_NaN(),
+    double risk_exposure_change = std::numeric_limits<double>::quiet_NaN()) {
   OrderUpdateBuilder builder_(_fbb);
-  builder_.add_risk_change(risk_change);
+  builder_.add_risk_exposure_change(risk_exposure_change);
+  builder_.add_risk_exposure(risk_exposure);
   builder_.add_sending_time_utc(sending_time_utc);
   builder_.add_last_traded_price(last_traded_price);
   builder_.add_last_traded_quantity(last_traded_quantity);
@@ -5802,7 +5824,8 @@ inline ::flatbuffers::Offset<OrderUpdate> CreateOrderUpdateDirect(
     const char *user = nullptr,
     int64_t sending_time_utc = 0,
     const char *client_order_id = nullptr,
-    double risk_change = std::numeric_limits<double>::quiet_NaN()) {
+    double risk_exposure = std::numeric_limits<double>::quiet_NaN(),
+    double risk_exposure_change = std::numeric_limits<double>::quiet_NaN()) {
   auto account__ = account ? _fbb.CreateString(account) : 0;
   auto exchange__ = exchange ? _fbb.CreateString(exchange) : 0;
   auto symbol__ = symbol ? _fbb.CreateString(symbol) : 0;
@@ -5849,7 +5872,8 @@ inline ::flatbuffers::Offset<OrderUpdate> CreateOrderUpdateDirect(
       user__,
       sending_time_utc,
       client_order_id__,
-      risk_change);
+      risk_exposure,
+      risk_exposure_change);
 }
 
 struct ParametersUpdate FLATBUFFERS_FINAL_CLASS : private ::flatbuffers::Table {