+++ include/roq/order_ack.hpp
@@ -43,7 +43,8 @@ struct ROQ_PUBLIC OrderAck final {
   std::string_view external_order_id;                //!< External order identifier
   std::string_view routing_id;                       //!< Routing identifier
   uint32_t version = {};                             //!< Version number (strictly increasing, optional)
-  double risk_change = NaN;                          //!< Risk change (cumulative >= remaining quantity)
+  double risk_exposure = NaN;                        //!< Risk exposure
+  double risk_exposure_change = NaN;                 //!< Risk exposure change
   double traded_quantity = NaN;                      //!< Quantity (total traded)
   std::chrono::nanoseconds round_trip_latency = {};  //!< Round-trip latency between gateway and exchange
 };
@@ -85,7 +86,8 @@ struct fmt::formatter<roq::OrderAck> {
         R"(external_order_id="{}", )"
         R"(routing_id="{}", )"
         R"(version={}, )"
-        R"(risk_change={}, )"
+        R"(risk_exposure={}, )"
+        R"(risk_exposure_change={}, )"
         R"(traded_quantity={}, )"
         R"(round_trip_latency={})"
         R"(}})"_cf,
@@ -105,7 +107,8 @@ struct fmt::formatter<roq::OrderAck> {
         value.external_order_id,
         value.routing_id,
         value.version,
-        value.risk_change,
+        value.risk_exposure,
+        value.risk_exposure_change,
         value.traded_quantity,
         value.round_trip_latency);
   }